Refactor PresentationView into smaller components
Breaks down the PresentationView component into smaller, more manageable sub-components for improved readability and maintainability.
