+++ src/pages/admin/surveys/campaigns/[id]/components/PresentationView/constants.ts
@@ -1,21 +1,12 @@
 
-export const COMPARISON_DIMENSIONS = [
-  'sbu',
+import { ComparisonDimension } from "./types/comparison";
+
+export const COMPARISON_DIMENSIONS: ComparisonDimension[] = [
+  'sbu', 
   'gender', 
-  'location',
+  'location', 
   'employment_type',
   'level',
   'employee_type',
   'employee_role'
-] as const;
-
-export const DIMENSION_TITLES = {
-  sbu: "Response Distribution by Department",
-  gender: "Response Distribution by Gender",
-  location: "Response Distribution by Location",
-  employment_type: "Response Distribution by Employment Type",
-  level: "Response Distribution by Level",
-  employee_type: "Response Distribution by Employee Type",
-  employee_role: "Response Distribution by Employee Role",
-  none: "No Comparison"
-} as const;
+];
