+++ src/pages/admin/surveys/campaigns/[id]/components/PresentationView/hooks/usePresentationNavigation.ts
@@ -1,0 +1,88 @@
+
+import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useToast } from "@/hooks/use-toast";
+
+interface UsePresentationNavigationProps {
+  id: string | undefined;
+  instanceId: string | null;
+  totalSlides: number;
+}
+
+export function usePresentationNavigation({ 
+  id, 
+  instanceId, 
+  totalSlides 
+}: UsePresentationNavigationProps) {
+  const navigate = useNavigate();
+  const { toast } = useToast();
+  const [currentSlide, setCurrentSlide] = useState(0);
+  const [isFullscreen, setIsFullscreen] = useState(false);
+
+  useEffect(() => {
+    if (!instanceId) {
+      toast({
+        title: "No instance selected",
+        description: "Please select an instance from the campaign page",
+        variant: "destructive",
+      });
+      navigate(`/admin/surveys/campaigns/${id}`);
+    }
+  }, [instanceId, id, navigate, toast]);
+
+  useEffect(() => {
+    const handleFullscreenChange = () => {
+      setIsFullscreen(!!document.fullscreenElement);
+    };
+    
+    document.addEventListener('fullscreenchange', handleFullscreenChange);
+    return () => document.removeEventListener('fullscreenchange', handleFullscreenChange);
+  }, []);
+
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === "ArrowLeft") {
+        setCurrentSlide((prev) => Math.max(0, prev - 1));
+      } else if (e.key === "ArrowRight") {
+        setCurrentSlide((prev) => Math.min(totalSlides - 1, prev + 1));
+      } else if (e.key === "f") {
+        toggleFullscreen();
+      }
+    };
+
+    document.addEventListener("keydown", handleKeyDown);
+    return () => document.removeEventListener("keydown", handleKeyDown);
+  }, [totalSlides]);
+
+  const toggleFullscreen = async () => {
+    try {
+      if (!document.fullscreenElement) {
+        await document.documentElement.requestFullscreen();
+      } else {
+        await document.exitFullscreen();
+      }
+    } catch (error) {
+      console.error("Error toggling fullscreen:", error);
+      toast({
+        title: "Fullscreen Error",
+        description: "Unable to toggle fullscreen mode. Please try again.",
+        variant: "destructive",
+      });
+    }
+  };
+
+  const handleBack = () => {
+    if (isFullscreen) {
+      document.exitFullscreen();
+    }
+    navigate(`/admin/surveys/campaigns/${id}`);
+  };
+
+  return {
+    currentSlide,
+    setCurrentSlide,
+    isFullscreen,
+    toggleFullscreen,
+    handleBack
+  };
+}
